Encapsulate enriched summary logic
