+++ src/main/java/hudson/plugins/gradle/enriched/ScanDetailService.java
@@ -1,0 +1,9 @@
+package hudson.plugins.gradle.enriched;
+
+import java.util.List;
+
+public interface ScanDetailService {
+
+  ScanDetail getScanDetail(String buildScanUrl);
+
+}
